Actividad8/requirements.txt
gradio
langchain
langchain-community
chromadb
python-dotenv
PyGithub
pypdf
langchain-google-genai
beautifulsoup4
markdown 